clearer separtion of synth and learn
